[修改] 乐优商城，goods页面，（添加）功能的第三步，根据规格参数查询，自己写的。 注意：第二步是它自己写好的集成了vue-Quill-Editor富文本编辑器

+++ leyou-item/leyou-item-service/src/main/java/com/leyou/item/mapper/CategoryBrandMapper.java
@@ -15,5 +15,5 @@ public interface CategoryBrandMapper extends Mapper<CategoryBrand> {
             "left join tb_brand tb\n" +
             "on tcb.brand_id=tb.id\n" +
             "where category_id=#{cid}")
-    List<Brand> queryBrandByCategoryID(Integer cid);
+    List<Brand> queryBrandByCategoryID(Long cid);
 }
